Added IFrame control
Added iframe control. basically a copy of Adam's iframe control, but
with some changes and cleanups to also make it work with the installer
program that's in this solution. It's also not using the textbox.

The internalcontrols are also moved into an internalcontrols namespace.
The installerprogram now sets nearly all the 'default' options. This has
to be improved with a property of somekind. Maybe have a property that
defines name, friendlyname, group and controltype methods via 1
attribute?