Updated silverlight control to also load the image
This is quite ugly, but for the silverlight control to load the image,
we need to pass it the url. There's a nice openFile() call that can be
used, but that URL for WHATEVER FREAKING SILVERLIGHT reason does not
work.

So, now we change/update the URL to something that does work (it
probably has to do with the query string parameters) <-- that's the ugly
part.

The reason this is so 'difficult' is that the silverlight upload control
in the OOTB component, actually does not show the image itself - it only
uploads.